+++ 11ty/src/_includes/layouts/base.njk
@@ -7,9 +7,8 @@
     {% else %}
       <title>{{ metadata.title }}</title>
     {% endif %}
-    {% if description %}
-      <meta name="description" content="{{ description }}" />
-    {% endif %}
+    <meta name="description" content="{{ description or metadata.description }}" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
     <link rel="alternate" type="application/rss+xml" title="Eli Gundry's Blog" href="/blog.rss" />
     <link rel="alternate" type="application/rss+xml" title="Eli Gundry's Feelings" href="/feelings.rss">
     <link rel="icon" href="data:image/svg+xml,<svg xmlns=%22http://www.w3.org/2000/svg%22 viewBox=%220 0 100 100%22><text y=%22.9em%22 font-size=%2290%22>{{ latestFeeling.emoji }}</text></svg>" />
@@ -17,8 +16,6 @@
     <link rel="preconnect" href="https://cdnjs.cloudflare.com" crossorigin="true" />
     <link rel="stylesheet" href="/css/style.css" />
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.2.0/css/all.min.css" />
-    <script src="/js/is-land.js" type="module"></script>
-    <script src="https://cdn.jsdelivr.net/npm/theme-change@2.0.2/index.js"></script>
     {% block head %}{% endblock %}
   </head>
   <body>
@@ -37,5 +34,8 @@
     {% endblock %}
 
     {% include 'fancyBackground.njk' %}
+
+    <script src="/js/is-land.js" type="module"></script>
+    <script src="https://cdn.jsdelivr.net/npm/theme-change@2.0.2/index.js"></script>
   </body>
 </html>
