+++ 11ty/src/_shortcodes/cachedImage.js
@@ -1,0 +1,10 @@
+const Image = require("@11ty/eleventy-img");
+
+module.exports = function (metadata, alt, sizes = "") {
+  return Image.generateHTML(metadata, {
+    alt,
+    sizes,
+    loading: "lazy",
+    decoding: "async",
+  });
+};
