+++ 11ty/src/_shortcodes/emojiText.js
@@ -1,0 +1,3 @@
+module.exports = function emojiText(emoji, label, text) {
+  return `<span role="img" aria-label="${label}">${emoji}</span> ${text}`;
+};
