+++ 11ty/src/_shortcodes/githubFileEmbed.js
@@ -1,0 +1,3 @@
+module.exports = async function (url) {
+  return "";
+};
