+++ 11ty/src/_shortcodes/googleSlideShow.js
@@ -1,0 +1,3 @@
+module.exports = function (url) {
+  return `<iframe src="${url}" frameborder="0" width="960" height="569" allowfullscreen="true"></iframe>`;
+};
