+++ 11ty/src/_shortcodes/index.js
@@ -1,0 +1,21 @@
+const cachedImage = require("./cachedImage");
+const cachedToolTipImage = require("./cachedToolTipImage");
+const emojiText = require("./emojiText");
+const feelingsActivityList = require("./feelingsActivityList");
+const formatDate = require("./formatDate");
+const githubFileEmbed = require("./githubFileEmbed");
+const googleSlideShow = require("./googleSlideShow");
+const image = require("./image");
+const resumeSkillList = require("./resumeSkillsList");
+
+module.exports = {
+  cachedImage,
+  cachedToolTipImage,
+  emojiText,
+  feelingsActivityList,
+  formatDate,
+  githubFileEmbed,
+  googleSlideShow,
+  image,
+  resumeSkillList,
+};
